--- junit5-cucumber-webdrivermanager/src/test/groovy/RunCucumberTest.groovy
@@ -1,5 +1,0 @@
-import io.cucumber.junit.platform.engine.Cucumber
-
-@Cucumber
-class RunCucumberTest {
-}
