+++ junit5-cucumber-webdrivermanager/src/test/groovy/sc/tyro/starter/junit5/RunCucumberTest.groovy
@@ -1,0 +1,17 @@
+package sc.tyro.starter.junit5
+
+import org.junit.platform.suite.api.ConfigurationParameter
+import org.junit.platform.suite.api.IncludeEngines
+import org.junit.platform.suite.api.SelectClasspathResource
+import org.junit.platform.suite.api.Suite
+
+import static io.cucumber.junit.platform.engine.Constants.PLUGIN_PROPERTY_NAME
+import static io.cucumber.junit.platform.engine.Constants.GLUE_PROPERTY_NAME
+
+@Suite
+@IncludeEngines("cucumber")
+@SelectClasspathResource("features")
+@ConfigurationParameter(key = PLUGIN_PROPERTY_NAME, value = "pretty")
+@ConfigurationParameter(key = GLUE_PROPERTY_NAME, value = "sc.tyro.starter.junit5.steps")
+class RunCucumberTest {
+}
