+++ junit5-cucumber-webdrivermanager/src/test/resources/junit-platform.properties
@@ -1,2 +1,1 @@
 cucumber.publish.quiet=true
-cucumber.glue=sc.tyro.starter.junit5
